fix: client builder checks if folder exists
adds use of script in README

+++ create-api-clients.sh
@@ -16,19 +16,24 @@ fi
 
 # Weather forecast API / yr
 # Downloads spec, adds host and scheme, creates client and adds it as an local dependency
-curl https://api.met.no/weatherapi/locationforecast/2.0/swagger | jq >>location-forecast-swagger.json
-echo "$(jq '. += {"host": "api.met.no", "schemes": ["https"]}' location-forecast-swagger.json)" >location-forecast-swagger-edit.json
-docker run --rm -v "${PWD}:/local" openapitools/openapi-generator-cli generate -i /local/location-forecast-swagger-edit.json -g rust -o /local/location-forecast-client --package-name=location-forecast-client --additional-properties=supportMiddleware=true
-cargo add --path ./location-forecast-client
-rm location-forecast-swagger.json location-forecast-swagger-edit.json
 
-curl https://docs.openrouteservice.org/all/docs | jq >>ors-swagger.json
-echo "$(jq 'del(.paths["/pois"])' ors-swagger.json)" >ors-swagger-edit.json
-docker run --rm -v "${PWD}:/local" openapitools/openapi-generator-cli generate -i /local/ors-swagger-edit.json -g rust -o /local/ors-client --package-name=ors-client --additional-properties=supportMiddleware=true
-cargo add --path ./ors-client
-rm ors-swagger.json ors-swagger-edit.json
-
-curl https://ws.geonorge.no/stedsnavn/v1/openapi.json | jq >>stedsnavn.json
-docker run --rm -v "${PWD}:/local" openapitools/openapi-generator-cli generate -i /local/stedsnavn.json -g rust -o /local/stedsnavn-client --package-name=stedsnavn-client --additional-properties=supportMiddleware=true
-cargo add --path ./stedsnavn-client
-rm stedsnavn.json
+if [ ! -d "./location-forecast-client/" ]; then
+  curl https://api.met.no/weatherapi/locationforecast/2.0/swagger | jq >>location-forecast-swagger.json
+  echo "$(jq '. += {"host": "api.met.no", "schemes": ["https"]}' location-forecast-swagger.json)" >location-forecast-swagger-edit.json
+  docker run --rm -v "${PWD}:/local" openapitools/openapi-generator-cli generate -i /local/location-forecast-swagger-edit.json -g rust -o /local/location-forecast-client --package-name=location-forecast-client --additional-properties=supportMiddleware=true
+  cargo add --path ./location-forecast-client
+  rm location-forecast-swagger.json location-forecast-swagger-edit.json
+fi
+if [ ! -d "./ors-client/" ]; then
+  curl https://docs.openrouteservice.org/all/docs | jq >>ors-swagger.json
+  echo "$(jq 'del(.paths["/pois"])' ors-swagger.json)" >ors-swagger-edit.json
+  docker run --rm -v "${PWD}:/local" openapitools/openapi-generator-cli generate -i /local/ors-swagger-edit.json -g rust -o /local/ors-client --package-name=ors-client --additional-properties=supportMiddleware=true
+  cargo add --path ./ors-client
+  rm ors-swagger.json ors-swagger-edit.json
+fi
+if [ ! -d "./stedsnavn-client/" ]; then
+  curl https://ws.geonorge.no/stedsnavn/v1/openapi.json | jq >>stedsnavn.json
+  docker run --rm -v "${PWD}:/local" openapitools/openapi-generator-cli generate -i /local/stedsnavn.json -g rust -o /local/stedsnavn-client --package-name=stedsnavn-client --additional-properties=supportMiddleware=true
+  cargo add --path ./stedsnavn-client
+  rm stedsnavn.json
+fi
